% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNG.R
\name{run_BNG}
\alias{run_BNG}
\title{Calculate BNG Units}
\usage{
run_BNG(polygons, lookup_path = NULL)
}
\arguments{
\item{polygons}{An sf object containing polygons. Must include a column "HabCode_B".}

\item{lookup_path}{Optional character path to the lookup CSV file. If not provided, the package's default file in inst/extdata is used.}
}
\value{
A numeric value representing the total BNG units.
}
\description{
This function calculates total Biodiversity Net Gain (BNG) units from an sf object of polygons by joining an inbuilt lookup
(using the column "HabCode_B") which supplies values for UKHab, Distinctiveness, condition and strategic_significance.
If the lookup does not include the columns "condition" or "strategic_significance", default values (3 and 1, respectively) are added.
}
\examples{
\dontrun{
  library(sf)
  # Assuming 'polys' is an sf object with a column "HabCode_B"
  total_units <- run_BNG(polys)
}
}
